+++ src/methods_stream.js
@@ -28,6 +28,10 @@ Atok.prototype.write = function (data) {
     this.needDrain = true
     return false
   }
+
+  // Check rules resolution
+  if (this._rulesToResolve) this._resolveRules()
+
   return this._tokenize()
 }
 /**
@@ -147,10 +151,6 @@ Atok.prototype._tokenize = function () {
       // Continue?
       } else if (p.continue !== null) {
         i += p.continue
-        if (i > this._rules.length || i < -1)
-          this._error( new Error('Out of bound rules index: ' + i + ' = ' +
-            (i - p.continue) + ' + ' + p.continue + ' > ' + this._rules.length
-          ))
         // Keep track of the rule index we are at
         this.ruleIndex = i + 1
         // Skip the token and keep going, unless rule returned 0
@@ -170,10 +170,6 @@ Atok.prototype._tokenize = function () {
       }
     } else if (p.continueOnFail !== null) {
       i += p.continueOnFail
-      if (i > this._rules.length || i < -1)
-        this._error( new Error('Out of bound rules index: ' + i + ' = ' +
-          (i - p.continueOnFail) + ' + ' + p.continueOnFail + ' > ' + this._rules.length
-        ))
       // Keep track of the rule index we are at
       this.ruleIndex = i + 1
     }